downloading files and senting to client - in process
