+++ backend/server.js
@@ -3,6 +3,7 @@ const cors = require('cors');
 const app = express()
 // app.set('view engine', 'ejs')
 const searchYoutube = require('./youtubeQuery');
+const youtubeDl = require('./youtubeDl');
 
 app.use(express.static('public'));
 app.use(cors())
@@ -18,4 +19,15 @@ app.get('/search/:text', async (req, res) => {
   console.log(`${response[0].snippet.title}\n-------------------------------`)
 })
 
+app.get('/download/:link/:title', async (req, res) => {
+  let link = req.params.link;
+  let title = req.params.title;
+  res.setHeader('Content-Disposition', 'attachment; filename=audio.mp3');
+  res.setHeader('Content-Type', 'audio/mpeg');
+  let outputPath = await youtubeDl(link, title);
+  res.download(outputPath);
+
+})
 app.listen(3000);
+
+// console.log(youtubeDl('https://www.youtube.com/watch?v=UBhs7CpKPSs', "Ordible test Audio.mp3"));
